\name{getArea}
\alias{getArea}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get peak area of a peak.
}
\description{
 Get peak area of a peak.
}
\usage{
getArea(eics, rtmin, rtmax)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{eics}{
EICs obtained by \code{getIsoEIC} function.
}
  \item{rtmin}{
Start of retention time.
}
  \item{rtmax}{
End of retention time.
}
}
