\name{getIsoSimi}
\alias{getIsoSimi}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get the similarity of two isotopic patterns
}
\description{
Get the similarity of two isotopic patterns
}
\usage{
getIsoSimi(iso_int1, iso_int2)
}

\arguments{
  \item{iso_int1}{
The detected isotopic pattern.
}
  \item{iso_int2}{
The real isotopic pattern.
}
}

\value{
a value from 0 to 1.
}