\name{LoadData}
\alias{LoadData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Load raw LC-MS data.
}
\description{
Load raw LC-MS data.
}
\usage{
LoadData(filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{The path of LC-MS data
}
}