\name{runTarMet}
\alias{runTarMet}
\title{
Run TarMat shiny application.
}
\description{
This is the only function in this package, which is used for calling the shiny app. Then you can process your data sets in the application.
}
\usage{
runTarMet()
}
\value{
A shiny app
}

\examples{
# not run
# runTarMet()
}