\name{getEIC}
\alias{getEIC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get EICs from raw data.
}
\description{
Get EICs from raw data.
}
\usage{
getEIC(raw, rtrange, mzrange)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{raw}{
Result obtain by \code{LoadData} 
}
  \item{rtrange}{
Range of retention time
}
  \item{mzrange}{
Range of m/z
}
}